games/sudoku/requirements.txt
typedb-client==2.16.1
